Restructure backend, add modular architecture and improve authentication
