+++ backend/package.json
@@ -1,27 +1,22 @@
 {
   "name": "codegrind-backend",
   "version": "1.0.0",
-  "description": "CodeGrind backend server",
-  "main": "server.js",
   "type": "module",
   "scripts": {
-    "dev": "nodemon server.js",
-    "start": "node server.js"
+    "start": "node src/server.js",
+    "dev": "nodemon src/server.js"
   },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
   "dependencies": {
-    "@prisma/client": "^5.22.0",
-    "bcryptjs": "^2.4.3",
+    "@prisma/client": "^5.0.0",
+    "bcrypt": "^5.1.0",
     "cors": "^2.8.5",
-    "dotenv": "^16.3.1",
+    "dotenv": "^16.0.3",
     "express": "^4.18.2",
-    "jsonwebtoken": "^9.0.2",
+    "jsonwebtoken": "^9.0.0",
     "leetcode-query": "^1.2.3"
   },
   "devDependencies": {
-    "nodemon": "^3.0.2",
-    "prisma": "^5.8.1"
+    "nodemon": "^3.1.9",
+    "prisma": "^5.0.0"
   }
 }
