+++ backend/routes/auth.js
@@ -1,77 +1,110 @@
 import { PrismaClient } from '@prisma/client';
-import bcrypt from 'bcryptjs';
+import bcrypt from 'bcrypt';
 import express from 'express';
 import jwt from 'jsonwebtoken';
 
 const router = express.Router();
 const prisma = new PrismaClient();
 
-// Register endpoint
-router.post('/register', async (req, res) => {
+// Login route
+router.post('/auth/login', async (req, res) => {
   try {
-    const { email, username, password } = req.body;
-    
+    const { username, password } = req.body;
+
+    // Find user by username or email
+    const user = await prisma.user.findFirst({
+      where: {
+        OR: [
+          { username },
+          { email: username }
+        ]
+      }
+    });
+
+    if (!user) {
+      return res.status(400).json({ message: 'User not found' });
+    }
+
+    // Verify password
+    const validPassword = await bcrypt.compare(password, user.password);
+    if (!validPassword) {
+      return res.status(400).json({ message: 'Invalid password' });
+    }
+
+    // Create token
+    const token = jwt.sign(
+      { userId: user.id },
+      process.env.JWT_SECRET || 'your-secret-key',
+      { expiresIn: '24h' }
+    );
+
+    // Send response
+    res.json({
+      token,
+      user: {
+        id: user.id,
+        username: user.username,
+        email: user.email
+      }
+    });
+  } catch (error) {
+    console.error('Login error:', error);
+    res.status(500).json({ message: 'Server error' });
+  }
+});
+
+// Register route
+router.post('/auth/register', async (req, res) => {
+  try {
+    const { username, email, password } = req.body;
+
     // Check if user already exists
     const existingUser = await prisma.user.findFirst({
       where: {
         OR: [
-          { email },
-          { username }
+          { username },
+          { email }
         ]
       }
     });
 
     if (existingUser) {
       return res.status(400).json({ 
-        error: 'User with this email or username already exists' 
+        message: 'Username or email already exists' 
       });
     }
 
     // Hash password
     const hashedPassword = await bcrypt.hash(password, 10);
-    
+
     // Create user
     const user = await prisma.user.create({
       data: {
-        email,
         username,
-        password: hashedPassword,
-      },
+        email,
+        password: hashedPassword
+      }
     });
 
-    // Generate JWT
+    // Create token
     const token = jwt.sign(
-      { userId: user.id }, 
-      process.env.JWT_SECRET,
+      { userId: user.id },
+      process.env.JWT_SECRET || 'your-secret-key',
       { expiresIn: '24h' }
     );
 
-    res.status(201).json({ token });
+    // Send response
+    res.status(201).json({
+      token,
+      user: {
+        id: user.id,
+        username: user.username,
+        email: user.email
+      }
+    });
   } catch (error) {
     console.error('Registration error:', error);
-    res.status(500).json({ error: 'Error creating user' });
-  }
-});
-
-// Login endpoint
-router.post('/login', async (req, res) => {
-  try {
-    const { email, password } = req.body;
-    const user = await prisma.user.findUnique({ where: { email } });
-    
-    if (!user) {
-      return res.status(401).json({ error: 'Invalid credentials' });
-    }
-
-    const validPassword = await bcrypt.compare(password, user.password);
-    if (!validPassword) {
-      return res.status(401).json({ error: 'Invalid credentials' });
-    }
-
-    const token = jwt.sign({ userId: user.id }, process.env.JWT_SECRET);
-    res.json({ token });
-  } catch (error) {
-    res.status(400).json({ error: error.message });
+    res.status(500).json({ message: 'Server error' });
   }
 });
