+++ backend/src/api/auth/auth.routes.js
@@ -1,0 +1,29 @@
+import express from 'express';
+import { authService } from './auth.service.js';
+
+const router = express.Router();
+
+// Login route
+router.post('/login', async (req, res) => {
+  try {
+    const { username, password } = req.body;
+    const result = await authService.login(username, password);
+    res.json(result);
+  } catch (error) {
+    console.error('Login error:', error);
+    res.status(400).json({ message: error.message });
+  }
+});
+
+// Register route
+router.post('/register', async (req, res) => {
+  try {
+    const result = await authService.register(req.body);
+    res.status(201).json(result);
+  } catch (error) {
+    console.error('Registration error:', error);
+    res.status(400).json({ message: error.message });
+  }
+});
+
+export default router;
