+++ backend/src/api/auth/auth.service.js
@@ -1,0 +1,74 @@
+import { PrismaClient } from '@prisma/client';
+import bcrypt from 'bcrypt';
+import jwt from 'jsonwebtoken';
+
+const prisma = new PrismaClient();
+
+export const authService = {
+  async login(username, password) {
+    const user = await prisma.user.findFirst({
+      where: {
+        OR: [
+          { username },
+          { email: username }
+        ]
+      }
+    });
+
+    if (!user) {
+      throw new Error('User not found');
+    }
+
+    const validPassword = await bcrypt.compare(password, user.password);
+    if (!validPassword) {
+      throw new Error('Invalid password');
+    }
+
+    const token = jwt.sign(
+      { userId: user.id },
+      process.env.JWT_SECRET || 'your-secret-key',
+      { expiresIn: '24h' }
+    );
+
+    return { token, user: this.sanitizeUser(user) };
+  },
+
+  async register(userData) {
+    const { username, email, password } = userData;
+
+    const existingUser = await prisma.user.findFirst({
+      where: {
+        OR: [
+          { username },
+          { email }
+        ]
+      }
+    });
+
+    if (existingUser) {
+      throw new Error('Username or email already exists');
+    }
+
+    const hashedPassword = await bcrypt.hash(password, 10);
+    const user = await prisma.user.create({
+      data: {
+        username,
+        email,
+        password: hashedPassword
+      }
+    });
+
+    const token = jwt.sign(
+      { userId: user.id },
+      process.env.JWT_SECRET || 'your-secret-key',
+      { expiresIn: '24h' }
+    );
+
+    return { token, user: this.sanitizeUser(user) };
+  },
+
+  sanitizeUser(user) {
+    const { password, ...sanitizedUser } = user;
+    return sanitizedUser;
+  }
+};
