+++ package.json
@@ -1,0 +1,7 @@
+{
+  "scripts": {
+    "start:backend": "cd backend && npm start",
+    "start:frontend": "cd codegrind-frontend && npm run dev",
+    "dev": "concurrently \"npm run start:backend\" \"npm run start:frontend\""
+  }
+}
